+++ src/Rask.Core/Components/Colgroup.cs
@@ -1,20 +1,6 @@
-using System.Globalization;
-using System.Text;
-
 namespace Rask.Core.Components;
 
-public sealed class Colgroup : Element
+public sealed class Colgroup : HtmlTableColElement
 {
     protected override string TagName => "colgroup";
-
-    public int? Span { get; set; }
-
-    protected override void WriteAttributes(StringBuilder sb)
-    {
-        base.WriteAttributes(sb);
-        if (Span is not null)
-        {
-            AppendAttr(sb, "span", Span.Value.ToString(CultureInfo.InvariantCulture));
-        }
-    }
 }