+++ src/Rask.Core/Components/H4.cs
@@ -1,6 +1,6 @@
 namespace Rask.Core.Components;
 
-public sealed class H4 : Element
+public sealed class H4 : HtmlHeadingElement
 {
     protected override string TagName => "h4";
 }