+++ src/Rask.Core/Components/H6.cs
@@ -1,6 +1,6 @@
 namespace Rask.Core.Components;
 
-public sealed class H6 : Element
+public sealed class H6 : HtmlHeadingElement
 {
     protected override string TagName => "h6";
 }