+++ src/Rask.Core/Components/Thead.cs
@@ -1,6 +1,6 @@
 namespace Rask.Core.Components;
 
-public sealed class Thead : Element
+public sealed class Thead : HtmlTableSectionElement
 {
     protected override string TagName => "thead";
 }